move route logic into app
